JimH patches to alleviate memory leaks associated with exec() calls. Specifically,
byteloader: Removed, this was serving as a cache for instances of this
class, but it's effect was to cause all classes loaded by
BytecodeLoader to never get garbage collected.  When making many
exec() calls (each of which creates a new Java class), this meant that
the memory for those classes stayed in use.  We now create a new
instance of BytecodeLoader for every class loaded.  This allows the
JVM to free the associated classes, and does not appear to impact
performance negatively (in fact, JimH's quick tests indicate that it
was actually faster)

makeClass(), makeCode(): Just return a new BytecodeLoader instance
every time.

clearLoader(): Removed, obsolete.

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%40830