PyJavaClass.addMethod(Method meth): Support doc-strings for java functions.
The algorithme is simply: whenever a java function is found on some
class, also looks for a public, static PyString called
"__doc__<function-name>". If such a field is found, the value of the
field is used as the doc-string for the function.


Example:

    public static PyString __doc__foo = new PyString(
        "foo(int) -> None\n"+
        "Do some fooness..."
    );

    public static void foo(int i) {
        ...
    }

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%401014