Support serialization of proxies.
When de-serializing an object graph, the utility class
PythonObjectInputStream must be used as the ObjectInputStream.

All proxies which implements java.io.Serializable are stored in a
global hashtable. When a request for a proxy class is detected
during de-serializing, the proxy name is looked for in the hashtable.

This means that the module that defines the classes must already have
been imported when de-serializing. THere are no automatic import of
the module that defines the classes.

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%401049