Augmented assignment. This adds python2.0 compatible augmented assignment. The new oprators are:
    +=
    -=
    *=
    /=
    %=
    &=
    |=
    ^=
    <<=
    >>=
    **=

In the syntax tree, the new operators become a subnode to expr_stmt. The
source:

    self.data *= n

creates the syntax tree:

    expr_stmt:null at line 68
     aug_multiply:null at line 68
      Dot_Op:null at line 68
       Name:self at line 68
       Name:data at line 68
      Name:n at line 68

In the dynamic code generation, the lhs of the assignment is visited
twice; the first time the mode==AUGGET, the second time with
mode=AUGSET. During the AUGGET fase, the final atom and the final
selector is saved (aload) to "local variables" in the JVM frame object
(do not confuse this with the python frame object). During the following
AUGSET fase, the local variable are restored (astore). The
identification of these "local variables" are stored in the node
(aug_tmp[1-4]).

In the static code generation, the lhs is also visited twice; first as a
normal visit, afterwards as a set(..) call. (See BaseEvaluator.
aug_binary_op). During both visits, the BaseEvaluator.AUG attributes
exists and contains the list of temporary assignment statement caused by
the final atom and the final selector. [This is a workaround for the
fact that the jpythonc design only allow for one expresion returned from
each syntax node. The "get" part of an augmented assignment can generate
up to 4 statements + 1 expression].

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%401086