Part of fix for "[ #462280 ] builtin method as a class variable".
I don't understand the reason for the code I removed with this change. If
there is a performance improvement hidden here, handling both the method
and the function aspect of PyBuiltinFunctionSet in the same way is surely
wrong.

Another hint of the strangeness is the invoke() with zero parameters
which does not check for PyBuiltinFunctionSet. At least now all 3
invoke() methods are working in a similar manner.

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%401826