Fix for [ #480390 ] main() does not throw exceptions.
Use a different property lists to the initProxy() and the runMain().
In initProxy() we still want to see java exception, in runMain()
it is now the JVM that prints the exception thrown from the main()
method.

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%401872