use the ouput generated from templates/make_binop.txt and introduce the needed helpers in PyObject and PyType.
Now we pass test_descr.subclass_right_op!  This solution is a compromise between speed and simplicity,
is still trying tough to make the fast path not too bad.

We can rethink this after 2.2.

Add a test to test_descr: subclass_cmp_right_op to check for the dispatch rules for comparisons,
this rule is more direct and less subtle than the general one. This test pass as is both
on top jython (with the fixes in _eq, _gt etc) and on top of Python 2.4, so we are reasonably
implementing the Python behavior as it stands today.

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%403017