Fix for http://bugs.jython.org/issue1264 'is not' test exhibits incorrect behaviour when wrapping Java objects
Included is a reshuffling of the patch test so that a regression test can pick
it up.  It isn't perfect since I had to use raw "assert" statements in the
python, but it is run by the Java unit test IdentityTest.

Thanks to Geoffrey French for the patch.

--HG--
extra : convert_revision : svn%3A79b11f02-f210-0410-99aa-fb189b3d1ac4/jython/trunk%406165