Use UTF-8 for file paths expressed in bytes.
This fairly extensive change regularises the approach to file and path names in
the interests of handling non-ascii paths correctly. See notes to issue #2356.
We are not finished with the consequential changes, but to commit work so far
helps make it manageable. regrtest runs with 24 failed tests.

--HG--
extra : amend_source : 397ad49b616f5ea95d4c9bda99fcfe800ae987b3