Enable JSON encoding of obj.__dict__ Fixes #2622
For a variety of historical reasons, the underlying object in Jython
for __dict__ is PyStringMap, not PyDictionary. Code that directly
works with this object in Java needs to account for this difference,
as is done now in this fix to the encoder.

--HG--
extra : amend_source : fa4bd85c6d72f0e5f7b536201f92e7b3ea3c3c17