Refactor PyType etc. to use a thread-safe non-blocking look-up.
This change re-works the apparatus behind PyType.fromClass and
addBuilder. It moves the static API of PyType to an inner "Registry",
where we use a ClassValue as a non-blocking cache in PyType look-up. The
construction of PyType and PyJavaType entries, bootstrapping of the type
system, and the initialisation of some core types have all been
re-thought, aiming for clearer logic. The "state journey" of a class and
its PyType from loading to exposure has been carefully considered.

The need to manage an explicit hard reference to exposed types has been
eliminated. The ClassValue classToType holds a plain (hard) reference,
and a re-worked test in test_jy_internals demonstrates that PyTypes are
correctly GC'd with the Java classes they represent to Python.