Respect default encoding when coercing str to unicode (addresses #2638).
This change corrects the implicit ascii or latin-1 assumption made when
accepting arguments in PyUnicode and PyString methods, adds tests to
test_unicode_jy, and makes small consequential changes to other modules.

The general effects is to allow, and decode, any byte buffer where a str
is acceptable. This is more liberal than CPython, except in __eq__ and
__ne__ which reproduce CPython limitations. It is not certain we have to
do that. Further change is needed to support all comparison operations.