Revise PyShadowString to speed up the target test and for simplicity.
This changes the order of tests in isTarget so regexes are compiled
once per test and to avoids the test if possible. Tests for equality and
startwith delegate to PyString rather than having their own
implementation. As a desirable by-product, we have
PyShadowString(a, b)[m,n] == PyShadowString(a[m,n], b[m,n])