Use relative path in launcher only if shorter than absolute path
We opted for relative paths to make paths and command-lines more
readable, and avoid some non-ascii encoding risks, when Jython is
installed locally to a user. The choice was counter-productive when
Jython was installed in a shared location.

--HG--
extra : amend_source : 8fd01b075f3387495b79d9b9ee5844638ff3f068