Determine console encoding without private method access (fixes #2659)
This change removes the reflective access to private method Console.encoding()
to which Java 9 objects, simplifing the determination of encoding a little.
We do not fall back to "file.encoding" as a guess, but on a hard-coded "utf-8".
Use of chcp was contributed by Oti Humbel. Use of "locale charmap" awaits
test on Unix-like platforms (check the Java property python.console.encoding).

Sub-process management is factored out of getSystemVersionString for shared use.