Re-work the publication of incomplete PyTypes (fixes #2834)
It is necessary to allow PyType.fromClass() to return incompletely
consructed types when building dependent PyTypes in a single thread.
This change intends they should not escape to other threads. We use a
RuntimeException to return a PyType without binding it to its class
prematurely.

Succinct debug methods are left in the code (disabled). We've had to
write these several times to address similar bugs.